I remove most of the files. Add new action listener foe the menu panel.
I removed most of the file because i don't know what they were used for. I also // the code the was using the removed files.  Also I added action listener to the menu panel. If any of the ice cream flavor is clicked the textbox will print. THE layout isn't done.